+++ lib/src/formula_supporters/formula_terms_completer.dart
@@ -66,4 +66,7 @@ class FormulaTermsCompleter extends FormulaTermsSupporter {
       }
     }
   }
+
+  @override
+  evaluate() => NullValue();
 }