+++ lib/src/operators/operators.dart
@@ -4,3 +4,4 @@ export 'multiply.dart';
 export 'division.dart';
 export 'power.dart';
 export 'compare.dart';
+export 'logical.dart';